db/pyproject.toml
[tool.poetry]
name = "training-tracking"
version = "0.1.0"
description = ""
authors = ["Michael Meissner"]
license = "MIT"
readme = "README.md"
packages = [{include = "training_tracking"}]

[tool.poetry.dependencies]
python = "^3.11"
yoyo-migrations = "^8.2.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
